requirements.txt
amqp==2.4.0
APScheduler==3.5.3
asn1crypto==0.24.0
billiard==3.5.0.5
blinker==1.4
CacheControl==0.12.5
cachetools==2.1.0
celery==4.2.1
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
cryptography==2.3.1
firebase-admin==2.15.0
Flask==1.0.2
Flask-JWT-Extended==3.15.0
Flask-SQLAlchemy==2.3.2
google-api-core==1.7.0
google-api-python-client==1.7.8
google-auth==1.6.2
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.3.0
google-cloud-core==0.29.1
google-cloud-firestore==0.31.0
google-cloud-storage==1.13.2
google-resumable-media==0.3.2
googleapis-common-protos==1.5.6
grpcio==1.18.0
httplib2==0.12.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
kombu==4.2.2.post1
MarkupSafe==1.1.0
msgpack==0.6.0
oauth2client==4.1.2
oauthlib==2.1.0
protobuf==3.6.1
psycopg2==2.7.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.3
pycparser==2.19
PyJWT==1.7.1
pyOpenSSL==18.0.0
PySocks==1.6.8
pytz==2018.9
redis==3.1.0
requests==2.21.0
requests-oauthlib==1.0.0
rq==0.13.0
rsa==3.4.2
simplejson==3.16.0
six==1.12.0
SQLAlchemy==1.2.0
tzlocal==1.5.1
uritemplate==3.0.0
urllib3==1.24.1
vine==1.2.0
Werkzeug==0.14.1
gunicorn